+++ Clockwork/Request/Log.php
@@ -1,20 +1,14 @@
 <?php
 namespace Clockwork\Request;
 
+use Psr\Log\AbstractLogger;
+use Psr\Log\LogLevel;
+
 /**
  * Data structure representing application log
  */
-class Log
+class Log extends AbstractLogger
 {
-	/**
-	 * Available log levels
-	 */
-	const DEBUG = 1;
-	const INFO = 2;
-	const NOTICE = 3;
-	const WARNING = 4;
-	const ERROR = 5;
-
 	/**
 	 * Array of log messages, with level and timestamp
 	 */
@@ -23,8 +17,20 @@ class Log
 	/**
 	 * Add a new timestamped message, with an optional level
 	 */
-	public function log($message, $level = Log::INFO)
+	public function log($level = LogLevel::INFO, $message, array $context = array())
 	{
+		if (is_object($message)) {
+			if (method_exists($message, '__toString')) {
+				$message = (string) $message;
+			} else if (method_exists($message, 'toArray')) {
+				$message = json_encode($message->toArray());
+			} else {
+				$message = json_encode((array) $message);
+			}
+		} else if (is_array($message)) {
+			$message = json_encode($message);
+		}
+
 		$this->data[] = array(
 			'message' => $message,
 			'level' => $level,