+++ Clockwork/Support/Monolog/Handler/ClockworkHandler.php
@@ -2,6 +2,7 @@
 namespace Clockwork\Support\Monolog\Handler;
 
 use Clockwork\Request\Log as ClockworkLog;
+
 use Monolog\Logger;
 use Monolog\Handler\AbstractProcessingHandler;
 
@@ -11,16 +12,6 @@ use Monolog\Handler\AbstractProcessingHandler;
 class ClockworkHandler extends AbstractProcessingHandler
 {
     protected $clockworkLog;
-    protected $levels = array(
-        Logger::DEBUG => ClockworkLog::DEBUG,
-        Logger::INFO => ClockworkLog::INFO,
-        Logger::NOTICE => ClockworkLog::NOTICE,
-        Logger::WARNING => ClockworkLog::WARNING,
-        Logger::ERROR => ClockworkLog::ERROR,
-        Logger::CRITICAL => ClockworkLog::ERROR,
-        Logger::ALERT => ClockworkLog::ERROR,
-        Logger::EMERGENCY => ClockworkLog::ERROR,
-    );
 
     public function __construct(ClockworkLog $clockworkLog)
     {
@@ -31,6 +22,6 @@ class ClockworkHandler extends AbstractProcessingHandler
 
     protected function write(array $record)
     {
-        $this->clockworkLog->log($record['message'], $this->levels[$record['level']]);
+        $this->clockworkLog->log($record['level'], $record['message']);
     }
 }