+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+dev
+NOTE: Clockwork\Request\Log::log method arguments have been changed from log($message, $level) to log($level, $message), levels are now specified via Psr\Log\LogLevel class, it's recommended to use shortcut methods for various levels (emergency, alert, critical,  error, warning, notice, info and debug($message))
+- clockwork log class now implements PSR logger interface, updated Laravel and Monolog support to use all available log levels
+- clockwork log now accepts objects and arrays as input and logs their json representation
+
 1.2
 - added support for Laravel 4.1
 - added facade for Laravel